+++ src/components/PlayArea/index.tsx
@@ -1,21 +1,24 @@
-import React, { useState } from 'react';
+import React from 'react';
 import Style from './style';
 import Card from '@components/Card';
 import Deck from '@components/Deck';
 
-const PlayArea = (): JSX.Element => {
-  const [active, setActive] = useState(-1);
+type PlayAreaProps = {
+  activeCard: number;
+  setActiveCard: React.Dispatch<React.SetStateAction<number>>;
+};
 
+const PlayArea = ({ activeCard, setActiveCard }: PlayAreaProps): JSX.Element => {
   const handleCardClick = (index: number) => {
-    if (index === active) index = -1;
-    setActive(index);
+    if (index === activeCard) index = -1;
+    setActiveCard(index);
   };
 
   const cards = [];
   for (let index = 0; index < 10; index++) {
     cards.push(
       <Card
-        active={active === index}
+        active={activeCard === index}
         index={index}
         key={index}
         handleCardClick={handleCardClick}