+++ src/components/PlayArea/spec.tsx
@@ -6,25 +6,20 @@ import Card from '@components/Card';
 import PlayArea from '@components/PlayArea';
 
 describe('PlayArea', () => {
+  const setActiveCard = jest.fn();
+  const PlayAreaMock = <PlayArea activeCard={-1} setActiveCard={setActiveCard} />;
+
   it('should render correctly', () => {
     const tree = renderer
-      .create(<PlayArea />)
+      .create(PlayAreaMock)
       .toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   describe('Cards', () => {
     it('should render the correct amount of Card components', () => {
-      const wrapper = mount(<PlayArea />);
+      const wrapper = mount(PlayAreaMock);
       expect(wrapper.find(Card)).toHaveLength(10);
-    });
-
-    it('should add/remove the .active class from Card components on mousedown', () => {
-      const wrapper = mount(<PlayArea />);
-      wrapper.find('.card').at(0).simulate('mousedown');
-      expect(wrapper.find('.card').at(0).hasClass('active')).toBe(true);
-      wrapper.find('.card').at(0).simulate('mousedown');
-      expect(wrapper.find('.card').at(0).hasClass('active')).toBe(false);
     });
   });
 });